Laget en klasse som holder på alle metoder brukt i programmet.
pickupFunctions inneholder metoder for å lage gui, finne pickuppoints i
tillegg til en enkel feilmeldingsmetode. Er også nytt at kunde får en
liten tilbakemelding hvis han/hun har gjort noe feil.
